refactor _google module for better naming
add new config file for templating
